fix(knitwit-99): fix structure

+++ src/main/java/com/knitwit/model/Notification.java
@@ -15,7 +15,7 @@ public class Notification {
     private int notificationId;
 
     @Column(name = "sender_id")
-    private Long senderId;
+    private int senderId;
 
     @Column(name = "title")
     private String title;
